Sets the Tyler lower court code on Metro South
Have seen several submissions to the Motion to Stay Eviction interview that
have the Metro South - Brockton court in production recently. This court
did not previously have a `tyler_lower_court_code` because it didn't exist
in the Tyler list (see 
https://github.com/SuffolkLITLab/docassemble-MACourts?tab=readme-ov-file#how-to-update-lower_court_code
for more information about this code).

In communications with the Trial Court since then about other interviews, they have said that

> the [Housing] Southeast Brockton Session [is] no longer accessible. Cases that were previously part of that division are now classified under Metro South.

Given that Tyler has not yet updated the lower court codes to accurately match the current state
of the court division, and that the Metro south court has been operating long enough to
expect appeals to begin appearing, I think it makes the most sense to use the
reverse of this statement, and set the Metro South courts to use the Housing Southeast code.

This is a production issue, so if we aren't sure this is the correct solution, we should reach
out to the Appeals court and directly ask what they think the solution should be.